+++ src/reader.rs
@@ -21,7 +21,7 @@ impl<R: Read> Reader<R> {
     pub fn read(mut self) -> Result<File> {
         let header = FileHeader::deserialize(&mut self.r)?;
         self.block_size = Some(header.block_size);
-        let mut sparse_file = File::new(header.block_size as usize);
+        let mut sparse_file = File::new(header.block_size);
 
         for _ in 0..header.total_chunks {
             self.read_chunk(&mut sparse_file)?;
@@ -33,11 +33,11 @@ impl<R: Read> Reader<R> {
     fn read_chunk(&mut self, sparse_file: &mut File) -> Result<()> {
         let header = ChunkHeader::deserialize(&mut self.r)?;
         let block_size = self.block_size.unwrap();
-        let size = (header.chunk_size * block_size) as usize;
+        let size = header.chunk_size * block_size;
 
         let chunk = match header.chunk_type {
             ChunkType::Raw => {
-                let mut buf = vec![0; size];
+                let mut buf = vec![0; size as usize];
                 self.r.read_exact(&mut buf)?;
                 Chunk::Raw { buf }
             }
